# api/openapi.yaml
openapi: 3.1.0
info:
  title: Processamento Distribuído de Apostas
  version: 1.0.0
  description: |
    Superfície HTTP do serviço: abertura e leitura de carteiras, ledger paginado por cursor,
    reconciliação, submissão e consulta de operações de provedores, health checks e métricas.

    Toda rota de negócio exige um access token OIDC do realm `wagering`. O escopo `wallets` é do
    serviço interno; o escopo `wagering` é dos provedores, que alcançam **apenas as próprias**
    operações — a identidade do token, e não o corpo, determina o `providerId` autorizado.

    **Como autenticar aqui:** clique em **Authorize**, preencha `client_id` e `client_secret` com
    uma das identidades abaixo, marque **só** o escopo dela e clique em *Authorize* de novo. O
    Swagger pede o token ao Keycloak (`client_credentials`) e o envia em toda chamada. Para trocar
    de identidade, *Logout* e repita.

    | client_id | client_secret | escopo | rotas |
    | --- | --- | --- | --- |
    | `internal-service` | `internal-service-secret` | `wallets` | carteiras, ledger, reconciliação |
    | `provider-a` | `provider-a-secret` | `wagering` | operações com `providerId: provider-a` |
    | `provider-b` | `provider-b-secret` | `wagering` | operações com `providerId: provider-b` |

    Respostas não-2xx são `application/problem+json` (RFC 9457). As classes de erro e os
    `failureCode` estáveis estão em `ARCHITECTURE.md`, seções 5 e 10; como subir o ambiente e obter
...
servers:
...
security:
  - oidc: [wagering]
...
paths:
...
  /wallets/{walletId}:
    get:
      tags: [wallets]
      summary: Lê uma carteira
      operationId: getWallet
      security:
        - oidc: [wallets]
      parameters:
        - $ref: '#/components/parameters/walletId'
        - $ref: '#/components/parameters/correlationId'
      responses:
        '200':
          description: A carteira
          headers:
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/WalletNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    oidc:
      type: oauth2
      description: |
        Keycloak, realm `wagering`, fluxo `client_credentials` — o enunciado põe o registro de senha
        e a emissão de tokens próprios fora de escopo. A API valida assinatura pelo JWKS, issuer,
        audiência e expiração; o `providerId` autorizado vem da claim `provider_id`, nunca do corpo.
        Identidades de teste e segredos locais estão no `README.md`.
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8081/realms/wagering/protocol/openid-connect/token
          scopes:
            wallets: Abrir, ler e reconciliar carteiras — serviço interno.
            wagering: Submeter e consultar as próprias operações — provedores.
  parameters:
    walletId:
      name: walletId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    correlationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Adotado quando enviado, gerado quando não, e sempre ecoado na resposta.
      schema: { type: string }
  headers:
    CorrelationId:
      description: O correlation id em uso, para o cliente citar em logs e suporte.
      schema: { type: string }
    RetryAfter:
      description: Segundos a esperar antes de reenviar.
      schema: { type: integer }
  schemas:
    Uuid:
      type: string
      format: uuid
      description: UUIDv7 gerado pela aplicação.
      example: 0192f291-27dd-7d3f-8071-5f8685deef37
    Money:
      type: object
      description: |
        Valor monetário exato — `int64` em unidades mínimas por dentro, string decimal no contrato,
        nunca ponto flutuante. Escala fixa de duas casas; na saída o sinal aparece, porque o
        `difference` da reconciliação é o armazenado menos o reconstruído e pode ser negativo.
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?[0-9]+\.[0-9]{2}$'
          example: "25.00"
        currency:
          $ref: '#/components/schemas/Currency'
      additionalProperties: false
...
    Currency:
      type: string
      description: |
        Código ISO 4217 em maiúsculas, restrito ao que o serviço liquida. Os fluxos do enunciado são
        em BRL; EUR e USD existem para os testes de moeda incompatível.
      enum: [BRL, EUR, USD]
    Wallet:
      type: object
      required: [id, playerId, balance, version]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        playerId: { $ref: '#/components/schemas/Uuid' }
        balance: { $ref: '#/components/schemas/Money' }
        version:
          type: integer
          format: int64
          description: Começa em `1` e sobe só quando o saldo muda.
          example: 1
...
    Problem:
      type: object
      description: Corpo de erro único da API (RFC 9457), servido como `application/problem+json`.
      required: [type, title, status, code]
      properties:
        type: { type: string, const: about:blank }
        title:
          type: string
          description: O texto do status HTTP.
          example: Bad Request
        status: { type: integer, example: 400 }
        code:
          type: string
          description: |
            Código estável da classe de erro. Numa rejeição de negócio é o `FailureCode` da
            transação; no que um construtor recusa antes de existir transação, é um dos códigos do
            contrato — `VALIDATION_FAILED`, `MALFORMED_BODY`, `WALLET_ALREADY_EXISTS`,
            `WALLET_NOT_FOUND`, `TRANSACTION_NOT_FOUND`, `IDEMPOTENCY_KEY_CONFLICT`,
            `EXTERNAL_TRANSACTION_CONFLICT`, `UNAUTHENTICATED`, `FORBIDDEN`, `SERVICE_UNAVAILABLE`
            ou `INTERNAL_ERROR`.
          example: VALIDATION_FAILED
        detail: { type: string }
        instance:
          type: string
          description: Numa rejeição de negócio, o caminho da transação em que a recusa foi gravada.
          example: /wagering/transactions/0192f298-345e-7e38-af88-e43f851a819d
        idempotentReplay:
          type: boolean
          description: Presente quando o problema é o desfecho gravado de uma submissão.
        errors:
          type: array
          description: Todas as violações do pedido de uma vez, para o cliente consertar numa ida só.
          items: { $ref: '#/components/schemas/Violation' }
    Violation:
      type: object
      required: [field, code, detail]
      properties:
        field:
          type: string
          description: O campo do corpo, o parâmetro de consulta ou o nome do header.
          example: money
        code:
          type: string
          enum: [REQUIRED, INVALID]
        detail: { type: string }
...
  responses:
    InvalidInput:
      description: Entrada inválida — `VALIDATION_FAILED` com `errors[]`, ou `MALFORMED_BODY`
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthenticated:
      description: Token ausente, malformado, expirado, com assinatura desconhecida ou audiência errada
      headers:
        WWW-Authenticate:
          schema: { type: string, const: Bearer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: about:blank
            title: Unauthorized
            status: 401
            code: UNAUTHENTICATED
            detail: the access token is not valid
    Forbidden:
      description: O token é válido, mas não carrega o escopo exigido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: about:blank
            title: Forbidden
            status: 403
            code: FORBIDDEN
            detail: the access token does not carry the wallets scope
    WalletNotFound:
      description: Carteira inexistente
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: about:blank
            title: Not Found
            status: 404
            code: WALLET_NOT_FOUND
            detail: wallet not found
...
    InternalError:
      description: Falha interna; a causa vai para o log, nunca para o corpo
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
            code: INTERNAL_ERROR
            detail: the request could not be completed
    Unavailable:
      description: |
        Indisponibilidade transitória — carteira disputada além do `DB_LOCK_TIMEOUT`, ou dependência
        inacessível. Nada foi aplicado, e o pedido pode ser reenviado com a mesma
        `Idempotency-Key`.
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: about:blank
            title: Service Unavailable
            status: 503
            code: SERVICE_UNAVAILABLE
            detail: the wallet is busy; the operation was not applied